LCAM-1032: Excluded ZipkinAutoConfiguration from crime commons auto-configuration
