+++ crime-commons-spring-boot-starter-rest-client/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,2 @@
 uk.gov.justice.laa.crime.commons.config.RestClientAutoConfiguration
+uk.gov.justice.laa.crime.commons.config.TraceAutoConfiguration
